Math-Tutor-Backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic>=2.10.0
requests>=2.31.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
sentence-transformers>=2.2.0
numpy>=1.24.0